add style to navigation bar
