+++ src/modules/menu.js
@@ -1,0 +1,6 @@
+const showMenu = () => {
+  const heading = document.createElement('h1');
+  heading.innerText = 'Menu goes here!';
+}
+
+export showMenu;
